+++ src/main/java/gift/exception/MemberExceptionHandler.java
@@ -21,7 +21,7 @@ public class MemberExceptionHandler {
     @ResponseStatus(HttpStatus.NOT_FOUND)
     @ExceptionHandler({LoginErrorException.class})
     public ErrorResult notFoundUserExHandle(LoginErrorException e) {
-        return new ErrorResult("로그인 에러", "아이디 또는 비밀번호가 일치하지 않습니다.");
+        return new ErrorResult("로그인 에러", e.getMessage());
     }
 
 }
