+++ src/main/java/gift/request/LoginMemberDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.request;
 
 public record LoginMemberDto(Long id) {
