+++ src/main/java/gift/request/WishListRequest.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.constraints.NotNull;
