+++ src/main/java/gift/resolver/LoginMemberArgumentResolver.java
@@ -1,7 +1,6 @@
-package gift.auth;
+package gift.resolver;
 
-import gift.dto.LoginMemberDto;
-import gift.model.Member;
+import gift.request.LoginMemberDto;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
