+++ src/main/java/gift/response/JoinResponse.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.response;
 
 public record JoinResponse(String email, String message) {
