+++ src/main/java/gift/response/ProductResponse.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.response;
 
 import gift.model.Product;
